+++ R/hy_daily_flows.R
@@ -131,9 +131,8 @@ hy_daily_flows <- function(station_number = NULL,
   colnames(dly_flows) <- c("STATION_NUMBER", "Date", "Parameter", "Value", "Symbol")
   
   
-  ## What stations were missed?
-  differ_msg(unique(stns), unique(dly_flows$STATION_NUMBER))
-  
-
-  dly_flows
+  attr(dly_flows, "missed_stns") <- setdiff(unique(stns), unique(dly_flows$STATION_NUMBER))
+  as.hy(dly_flows)
 }
+
+