+++ R/hy_plot.R
@@ -10,7 +10,91 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and limitations under the License.
 
-#' Convenience function to plot realtime data
+
+#' plot hy object
+#' 
+#' This method plots daily time series data from HYDAT. Functionality is very basic
+#' and is intended for exploratory purposes only.
+#' 
+#' @param x Object created by either a hy_daily_* or realtime_dd data retrieval function
+#' @param ... passed to \code{plot}
+#' 
+#' @method plot hy
+#' @name plot
+#' 
+#' @examples 
+#' \dontrun{
+#' # One station
+#' fraser <- hy_daily_flows("08MF005")
+#' plot(fraser)
+#' }
+#' 
+#' @export
+#' 
+plot.hy <- function(x = NULL, ...){
+  if(!all(c("STATION_NUMBER", "Date", "Parameter", "Value") %in% names(x))){
+    stop("plot methods only currently accept daily values", call. = FALSE)
+  }
+  
+  ### Join with meta data to get station name
+  hydf <- dplyr::left_join(x, 
+                           suppressMessages(hy_stations()), 
+                           by = c("STATION_NUMBER"))
+  
+  hydf$STATION <- paste(hydf$STATION_NAME, hydf$STATION_NUMBER, sep = " - ")
+  
+  hydf$STATION <- factor(hydf$STATION)
+  
+  num_stns <- length(unique(hydf$STATION))
+  
+  if(num_stns > 4) stop("You are trying to plot more than four stations at once.", call. = FALSE)
+  
+  if(num_stns > 2){
+    m <- matrix(c(1,1,2,3,4,5,6,6),nrow = 4,ncol = 2,byrow = TRUE)
+    graphics::layout(mat = m,heights = c(0.1,0.35,0.35,0.2))
+  } 
+  
+  if(num_stns == 2){
+    m <- matrix(c(1,1,2,3,4,4),nrow = 3,ncol = 2,byrow = TRUE)
+    graphics::layout(mat = m,heights = c(0.2,0.6,0.2))
+  } 
+  
+  if(num_stns == 1){
+    m <- matrix(c(1,2,3),nrow = 3,ncol = 1,byrow = TRUE)
+    graphics::layout(mat = m,heights = c(0.2,0.6,0.2))
+  } 
+  
+  graphics::par(mar=c(1,1,1,1))
+  graphics::plot.new()
+  graphics::text(0.5,0.5,"Historical Water Survey of Canada Gauges",cex=2,font=2)
+  
+  for(i in seq_along(unique(hydf$STATION))){
+    graphics::par(mar = c(2,2,1,1))
+    graphics::plot(Value ~ Date,
+                   data = hydf[hydf$STATION == unique(hydf$STATION)[i],],
+                   xlab="Date", 
+                   bty= "L",
+                   pch = 20, 
+                   cex = 1,
+                   ...)
+
+    graphics::title(main=paste0(unique(hydf$STATION)[i]), cex.main = 1.75)
+    
+  }
+  
+  
+  graphics::plot(1, type = "n", axes=FALSE, xlab="", ylab="")
+  # graphics::legend(x = "top", inset = 0,
+  #                  legend = unique(hydf$STATION), 
+  #                  fill = unique(hydf$STATION),
+  #                  bty = "n",
+  #                  cex = 1, horiz = TRUE)
+  
+  invisible(TRUE)
+  
+}
+
+#' This function is deprecated in favour of generic plot methods
 #' 
 #' This is an easy way to visualize a single station using base R graphics. 
 #' More complicated plotting needs should consider using \code{ggplot2}. Inputting more 
@@ -20,20 +104,10 @@
 #' @param station_number A (or several) seven digit Water Survey of Canada station number. 
 #' @param Parameter Parameter of interest. Either "Flow" or "Level".
 #' 
-#' @return A plot of recent realtime values
-#' 
-#' @examples 
-#' \dontrun{
-#' ## One station
-#' hy_plot("08MF005")
-#' 
-#' ## Multiple stations
-#' hy_plot(c("07EC002","01AD003"))
-#' }
-#' 
 #' @export
-
 hy_plot <- function(station_number = NULL, Parameter = c("Flow","Level", "Suscon","Load")){
+  message("hy_plot has been deprecated in favour of using the generic R plot method and will disappear in future versions.")
+  
   
   Parameter <- match.arg(Parameter, several.ok = TRUE)
   
@@ -59,7 +133,7 @@ hy_plot <- function(station_number = NULL, Parameter = c("Flow","Level", "Suscon
   #palette(rainbow(length(unique(rldf$STATION_NUMBER))))
   
   if(length(params) > 2){
-  #par(mfrow = c(2, 2))
+    #par(mfrow = c(2, 2))
     m <- matrix(c(1,1,2,3,4,5,6,6),nrow = 4,ncol = 2,byrow = TRUE)
     
     graphics::layout(mat = m,heights = c(0.1,0.35,0.35,0.2))
@@ -77,7 +151,7 @@ hy_plot <- function(station_number = NULL, Parameter = c("Flow","Level", "Suscon
     
     graphics::layout(mat = m,heights = c(0.2,0.6,0.2))
   } 
-
+  
   graphics::par(mar=c(1,1,1,1))
   graphics::plot.new()
   #graphics::plot(1, type = "n", axes=FALSE, xlab="", ylab="")
@@ -86,16 +160,16 @@ hy_plot <- function(station_number = NULL, Parameter = c("Flow","Level", "Suscon
   for(i in seq_along(params)){
     graphics::par(mar = c(2,2,1,1))
     graphics::plot(Value ~ Date,
-                 data = hydf[hydf$Parameter == params[i],],
-                 col = hydf$STATION,
-                 xlab="Date", 
-                 ylab = paste0(params[i]),
-                 bty= "L",
-                 pch = 20, cex = 1)
+                   data = hydf[hydf$Parameter == params[i],],
+                   col = hydf$STATION,
+                   xlab="Date", 
+                   ylab = paste0(params[i]),
+                   bty= "L",
+                   pch = 20, cex = 1)
     
     graphics::title(main=paste0(params[i]), cex.main = 1.75)
-  
-  
+    
+    
   }
   
   
@@ -105,6 +179,4 @@ hy_plot <- function(station_number = NULL, Parameter = c("Flow","Level", "Suscon
                    fill = unique(hydf$STATION),
                    bty = "n",
                    cex = 1, horiz = TRUE)
-  
-  
 }