+++ tests/testthat/test_station_choice.R
@@ -45,3 +45,8 @@ test_that("province in any case is accepted",{
   expect_silent(stns <- tidyhydat:::station_choice(hydat_con, station_number = NULL, prov_terr_state_loc = prov))
   expect_identical(toupper(prov), unique(hy_stations(hydat_path, station_number = stns)$PROV_TERR_STATE_LOC))
 })
+
+test_that("'CA' to prov_terr_state_loc argument returns only Canadian stations",{
+  only_canada <- unique(hy_stations(prov_terr_state_loc = "CA", hydat_path = hy_test_db())$PROV_TERR_STATE_LOC)
+  expect_equal(c("NB", "AB", "YT", "BC"), only_canada)
+})